requirements.txt
apt-xapian-index==0.49
bmp280==1.0.0
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
configobj==5.0.6
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
distro-info==1.0
i2cdevice==1.0.0
idna==2.10
iotop==0.6
louis==3.16.0
mqtt==0.0.1
netifaces==0.10.9
paho-mqtt==2.1.0
pexpect==4.8.0
Pillow==8.1.2
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
python-apt==2.2.1
python-debian==0.1.39
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.59
requests==2.25.1
six==1.16.0
smbus==1.1.post2
smbus2==0.5.0
terminator==2.1.0
unattended-upgrades==0.1
urllib3==1.26.5
wiringpi==2.60.1
xdg==5
